+++ zfit/core/parameter.py
@@ -20,6 +20,8 @@ from zfit.util.container import convert_to_container
 from . import interfaces as zinterfaces
 from .interfaces import ZfitModel, ZfitParameter, ZfitIndependentParameter
 from ..core.baseobject import BaseNumeric
+from ..minimizers.fitresult import FitResult
+from ..minimizers.interface import ZfitResult
 from ..settings import ztypes, run
 from ..util import ztyping
 from ..util.cache import invalidate_graph
@@ -839,7 +841,7 @@ def convert_to_parameter(value, name=None, prefer_constant=True, dependents=None
 
 def set_values(params: Union[Parameter, Iterable[Parameter]],
                values: Union[ztyping.NumericalScalarType,
-                             Iterable[ztyping.NumericalScalarType]]):
+                             Iterable[ztyping.NumericalScalarType], ZfitResult]):
     """Set the values (using a context manager or not) of multiple parameters.
 
     Args:
@@ -850,7 +852,14 @@ def set_values(params: Union[Parameter, Iterable[Parameter]],
 
     """
     params = convert_to_container(params)
-    if len(params) > 1:
+    if isinstance(values, ZfitResult):
+        result = values
+        values = []
+        for param in params:
+            if not param in result.params:
+                raise ValueError(f"Cannot set {param} with {repr(FitResult)} as it is not contained.")
+            values.append(result.params[param]['value'])
+    elif len(params) > 1:
         if not tf.is_tensor(values) or isinstance(values, np.ndarray):
             values = convert_to_container(values)
             if not len(params) == len(values):